Turned _updateStars into an observer;
This change will ensure that `_updateStare` is an observer, and any change to `rate` will imply a real-time change to the rendered stars. This means that `rate` can be changed dynamically. Note that I took out references to `this.rate` within `_updateStars` (since I am using the observer's parameter) and made sure to pass `rate` in the functions that use it.